Frontend restructured
