+++ frontend/src/Components/iconButton/IconButton.tsx
@@ -1,7 +1,7 @@
 import withStyles from "@material-ui/core/styles/withStyles";
 import * as React from "react";
 
-export interface ButtonProps {
+export interface IconButtonProps {
   classes?: any;
   icon?: string;
   text?: string;
@@ -22,14 +22,13 @@ const styles = withStyles((theme) => ({
   }
 }));
 
-class Button extends React.Component<ButtonProps, any> {
-  constructor(props: ButtonProps) {
+class IconButton extends React.Component<IconButtonProps, any> {
+  constructor(props: IconButtonProps) {
     super(props);
   }
   render() {
     const { classes, icon, text } = this.props;
     return (
-      // <div className={classes.container} onClick={()=>alert('Hello world!')}>
       <div className={classes.container}>
         <i className={icon}></i>
         <span className={classes.font}>{text}</span>
@@ -38,4 +37,4 @@ class Button extends React.Component<ButtonProps, any> {
   }
 }
 
-export default styles(Button);
+export default styles(IconButton);
