+++ frontend/src/pages/receivedSignature/ReceivedSignature.tsx
@@ -1,6 +1,6 @@
 import { withStyles } from "@material-ui/core";
 import React, { Component } from "react";
-import DetailedSign from "../../DetailedSign/DetailedSign";
+import DetailedSign from "../../components/detailedSign/DetailedSign";
 
 const styles = withStyles({
   container: {
